[DEV-1227] Replace pyyaml
Unlike PyYAML, ruamel.yaml supports:
* YAML <= 1.2. PyYAML only supports YAML <= 1.1
This is vital, as YAML 1.2 intentionally breaks backward compatibility
with YAML 1.1 in several edge cases. This would usually be a bad thing.
In this case, this renders YAML 1.2 a strict superset of JSON. Since
YAML 1.1 is not a strict superset of JSON, this is a good thing.
* Roundtrip preservation
When calling yaml.dump() to dump a dictionary loaded by a prior call
to yaml.load():

See more details at https://yaml.readthedocs.io/en/latest/pyyaml.html